chore(aws_s3 sink): Cleanup files and rename tests module
