+++ src/sinks/aws_s3/mod.rs
@@ -1,7 +1,6 @@
 mod config;
 mod sink;
 
-#[cfg(test)]
-mod tests;
+mod integration_tests;
 
 pub use self::config::S3SinkConfig;
